+++ react/src/components/context/UserContext.js
@@ -4,7 +4,7 @@ export const UsersContext = createContext();
 
 export const UsersProvider = ({ children }) => {
   const [users, setUsers] = useState([]);
-  const port = 3001;
+  const port = 3002;
   const fetchUsers = async () => {
     const response = await fetch(`http://localhost:${port}/users`);
     const data = await response.json();
@@ -13,22 +13,34 @@ export const UsersProvider = ({ children }) => {
   };
 
   const deleteUser = async (userId) => {
-    try {
-      await fetch(`http://localhost:${port}/users/${userId}`, {
-        method: "DELETE",
-      });
-      setUsers(
-        users.filter((user) => {
-          if (user._id) {
-            return user._id !== userId;
-          } else if (user.id) {
-            return user.id !== userId;
+    const password = prompt("Enter the delete password:");
+    if (password) {
+      try {
+        const response = await fetch(
+          `http://localhost:${port}/users/${userId}`,
+          {
+            method: "DELETE",
+            headers: {
+              "Content-Type": "application/json",
+            },
+            body: JSON.stringify({ password }),
           }
-          return false;
-        })
-      );
-    } catch (error) {
-      console.error("Error deleting user:", error);
+        );
+        if (response.status === 204) {
+          setUsers(
+            users.filter((user) => {
+              if (user._id) {
+                return user._id !== userId;
+              } else if (user.id) {
+                return user.id !== userId;
+              }
+              return false;
+            })
+          );
+        }
+      } catch (error) {
+        console.error("Error deleting user:", error);
+      }
     }
   };
